pyproject.toml
[project]
name = "data-analyst"
version = "0.1.0"
description = "Data-analyst agent framework"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
    "fastapi>=0.128.0",
    "matplotlib>=3.10.8",
    "openai>=2.15.0",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "seeds-clients",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
]

[tool.uv]
package = false

[tool.uv.sources]
seeds-clients = { git = "https://github.com/leokeba/seeds-clients" }

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]
